+++ game/models.py
@@ -4,6 +4,10 @@ class Tile:
     def __init__(self, letter, value):
         self.letter = letter
         self.value = value
+
+    def __str__(self):
+        return self.letter
+
     def __repr__(self):
         return f"{self.letter}:{self.value}"